Update README.md
day1

1.Importing the scanner class to take user input. 
2.Creating an object to access scanner class. 
3.considering the variables to store item details. 
4.printing the details of each item to the user using a while loop. 5.Switch case is used to calculate the price of each item selected by the user.

day2_mathOperations

A calculate function is created to calculate all the operations between two operands. As we have to return all the calculations we should return an array [] of the results. In the main function we should call the calculate function by passing the two parameters and retrieving the results from the array through indexing.